add auto build hook

+++ kodi/build-afpfs-ng.sh
@@ -125,32 +125,44 @@ main()
 		cd "$HOME"
 	fi
 	
-	# inform user of packages
-	echo -e "\n############################################################"
-	echo -e "If package was built without errors you will see it below."
-	echo -e "If you don't, please check build dependcy errors listed above."
-	echo -e "############################################################\n"
+	# If "build_all" is requested, skip user interaction
 	
-	echo -e "Showing contents of: ${git_dir}/build: \n"
-	ls "${git_dir}/build" | grep -E *.deb
-
-	echo -e "\n==> Would you like to transfer any packages that were built? [y/n]"
-	sleep 0.5s
-	# capture command
-	read -ep "Choice: " transfer_choice
+	if [[ "$build_all == "yes" ]]; then
 	
-	if [[ "$transfer_choice" == "y" ]]; then
-	
-		# cut files
-		if [[ -d "${git_dir}/build" ]]; then
-			scp ${git_dir}/build/*.deb mikeyd@archboxmtd:/home/mikeyd/packaging/SteamOS-Tools/incoming
-
-		fi
+		echo -e "\n==INFO==\nAuto-build requested"
+		mv ${git_dir}/build/*.deb "$auto_build_dir"
+		sleep 2s
 		
-	elif [[ "$transfer_choice" == "n" ]]; then
-		echo -e "Upload not requested\n"
-	fi
+	else
+		
+		# inform user of packages
+		echo -e "\n############################################################"
+		echo -e "If package was built without errors you will see it below."
+		echo -e "If you don't, please check build dependcy errors listed above."
+		echo -e "############################################################\n"
+		
+		echo -e "Showing contents of: ${git_dir}/build: \n"
+		ls "${git_dir}/build" | grep -E *.deb
+	
+		echo -e "\n==> Would you like to transfer any packages that were built? [y/n]"
+		sleep 0.5s
+		# capture command
+		read -ep "Choice: " transfer_choice
+		
+		if [[ "$transfer_choice" == "y" ]]; then
+		
+			# cut files
+			if [[ -d "${git_dir}/build" ]]; then
+				scp ${git_dir}/build/*.deb mikeyd@archboxmtd:/home/mikeyd/packaging/SteamOS-Tools/incoming
+	
+			fi
+			
+		elif [[ "$transfer_choice" == "n" ]]; then
+			echo -e "Upload not requested\n"
+		fi
 
+	fi
+	
 }
 
 # start main
